+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## vNext
 - Added support for ReSharper and Rider 2024.3
+- Fixed default settings marker
 
 ## 2024.2.0
 - Added support for ReSharper and Rider 2024.2